Adding some boyes

+++ src/model/Character.java
@@ -2,14 +2,14 @@ package model;
 
 public abstract class Character extends AbstractEntity {
 
-	private int health;
+	private double health;
 
 	public Character(AABB aabb, int depth, double mass, Level level) {
 		super(aabb, depth, mass, level);
 		this.health = this.getMaxHealth();
 	}
 
-	public int getHealth(){ return health; }
+	public double getHealth(){ return health; }
 	public int getMaxHealth(){ return 100;}
-	public void damage(int damage){ this.health -= damage; }
+	public void damage(double damage){ this.health -= damage; }
 }
